discord_bot/requirements.txt
py-cord==2.6.0
httpx==0.27.2
python-dotenv==1.0.1
